+++ pos_longpolling/static/src/js/PosConnection.js
@@ -1,8 +1,8 @@
-odoo.define('pos_longpolling.connection', function(require){
+odoo.define('pos_longpolling.PosConnection', function(require){
     "use strict";
 
-    var LongpollingModel = require('pos_longpolling.model');
-    var bus_longpolling = require('pos_longpolling.bus_longpolling');
+    var LongpollingModel = require('pos_longpolling.LongpollingModel');
+    var bus_longpolling = require('pos_longpolling.LongpollingBus');
     var core = require('web.core');
     var SyncBusService = require('pos_longpolling.SyncBusService');
     var crash_manager = require('web.crash_manager');