+++ pos_multi_session_restaurant/__manifest__.py
@@ -8,7 +8,7 @@
     "category": "Point of Sale",
     # "live_test_url": 'http://apps.it-projects.info/shop/product/pos-multi-session?version=12.0',
     "images": ['images/s2.png'],
-    "version": "12.0.3.3.1",
+    "version": "12.0.3.3.2",
     "application": False,
 
     "author": "IT-Projects LLC, Ivan Yelizariev",