+++ pos_multi_session_sync/__manifest__.py
@@ -6,8 +6,8 @@
     "summary": """Sync orders via main or separate (e.g. local) server""",
     "category": "Point of Sale",
     # "live_test_url": 'http://apps.it-projects.info/shop/product/pos-multi-session?version=12.0',
-    "images": [],
-    "version": "12.0.1.0.6",
+    "images": ['images/pos_multi_session_sync.jpg'],
+    "version": "12.0.1.0.7",
     "application": False,
 
     "author": "IT-Projects LLC, Kolushov Alexandr",