+++ pos_order_cancel_restaurant/static/src/js/tour.js
@@ -1,3 +1,7 @@
+/* Copyright 2017-2018 Ivan Yelizariev <https://it-projects.info/team/yelizariev>
+ * Copyright 2017-2018 Gabbasov Dinar <https://it-projects.info/team/GabbasovDinar>
+ * Copyright 2018-2019 Kolushov Alexandr <https://it-projects.info/team/KolushovAlexandr>
+ * License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html). */
 odoo.define('pos_order_cancel_restaurant.tour', function(require) {
     "use strict";