+++ pos_multi_session/doc/changelog.rst
@@ -6,6 +6,7 @@
 `4.2.9`
 -------
 
+- **Fix:** Error "journal_id is not defined" after add new journal to another POS with same multi-session
 - **Fix:** Expected singleton error on adding POSes to a multisession
 
 `4.2.8`